+++ hyperspace/ConsumerAbstractAdaptor.h
@@ -68,6 +68,7 @@ class ConsumerAbstractAdaptor : public AbstractWaveTarget
         bool payloadToValue(const QByteArray &payload, double *value);
         bool payloadToValue(const QByteArray &payload, QString *value);
         bool payloadToValue(const QByteArray &payload, QDateTime *value);
+        bool payloadToValue(const QByteArray &payload, QList<QVariant> *value);
 
     private:
         class Private;